+++ nixos-modules/gateway-vm.nix
@@ -97,6 +97,7 @@ let
               image
               diskName
               managedIdentityResourceId
+              managedIdentityClientId
               cpu
               memory
               autoSuspendIntervalSecs